changed the gpt code so it makes sense atleast- NOT TESTED
