+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeamCore/Drive/Implementation/LocalPathFetcher.java
@@ -1,9 +1,15 @@
-package org.firstinspires.ftc.teamcode.TeamCore.Pathing;
+package org.firstinspires.ftc.teamcode.TeamCore.Drive.Implementation;
+
+import org.firstinspires.ftc.teamcode.TeamCore.Drive.PathFetcher;
+import org.firstinspires.ftc.teamcode.TeamCore.Pathing.BasePath;
+import org.firstinspires.ftc.teamcode.TeamCore.Pathing.LinearPath;
+import org.firstinspires.ftc.teamcode.TeamCore.Pathing.Point;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import EngineCore.EngineCore;
+import EngineCore.TestingEnviromentCore;
 
 /**
  * Default PathFetcher implementation that returns locally stored paths
@@ -25,12 +31,26 @@ public class LocalPathFetcher extends PathFetcher {
         this(true, core);
     }
 
-    @Override
     public BasePath fetch(Point start, Point end) {
         return paths.get(key(start, end));
     }
 
     private String key(Point a, Point b){
         return a.cartesianX + "," + a.cartesianY + "_" + b.cartesianX + "," + b.cartesianY;
+    }
+
+    @Override
+    protected void step(EngineCore core) {
+
+    }
+
+    @Override
+    protected void update(EngineCore core) {
+
+    }
+
+    @Override
+    protected int test(TestingEnviromentCore core) {
+        return 0;
     }
 }
